Use Raft with up to 20 compute nodes (Test only)
binaries currently ignore RAFT.

With the network issues addressed partially, we can now establish
all the needed connection ensuring that we are only doing it once per pair.

The other aspect is the cloning needed on every spawned tasks.
this is a problem because the task modify the current block and we
want to then use it, so we need to use the same instance.

2 options:
 - Not clone (used here) extracting the nodes and re-adding them.
 - Not clone with Arc Mutext.

The current solution is very basic and should be cleaned up.
Eventually we probably want to make ComputeNode, MinerNode etc non clonable, and only clone Node or any thing that needs to (see compute.rs:raft_loop for an example)

Test:
cargo test
cargo build