Swiitch network test to use lock instead of clone.
Clonning our instances duplicate a lot of information, but also,
we do not want to spawn different object for running the tasks as
we want to update a single instance with all the modifications.

Each test need to be careful to avoid dead lock, however that should
be fairly reasonable and simpler to write than the alternative that would
use the take_compute/add_back_compute pattern. The result should be
cleaner and easier to decompose tests.


Test:
Cargo test