Use timeout for Raft generate_block
Add initial timeout (100ms instead of 1sec for tests), so we do not
require explicit call to propose a new block.

Note:
Only generate the proposal on a single node otherwise the test
generate some unwrap panic on sub task because we stop running the
test before all proposals are done (i.e once the first block proposal is commited).

Note:
bin/compute does not work as it call the trait function directly and so
the call to propose_local_transactions does not happen.

Test:
cargo test