Maintain in flight tx_pool len and items
In order to check against the tx_pool limit, we keep track of the ones
that have been voted, but not yet commited.

It is also possible that a proposal was dropped, we keep track of the
ones in flight so we may resend them later (i.e if leader dropped).

Test:
new unit test
cargo test