Ensure connect_to complete if other node is full
We were waiting for the notification to complete the `connect_to` funcition.
But if the peer decided not to add us, we would drop the peer without notifying here.

Use oneshot channel instead of Notify to make sure that receiver is always notified
whether it is because we receive the handshake response, or dropped the peer.

Test:
cargo clippy
cargo test
src/bin/node_settings_local_raft_run.sh