Refactor test_utils to ease removing/adding nodes
We want to test killing/restarting a node. So we need to stop it drop it, and then
add a new node in its place.

To ease that combine the 4 node collection in a single one and extract funcitons
to start/stop nodes so they can be used for a single node.

Test:
cargo clippy
cargo test