Fix lock expiry
we were using the current_mined_block, however, this is not consensused info.
It is only available when miners have not yet found a POW.
It would be a false negative on lock expiry if current_mined_block was none.

Fix code, and simplify it, removing the subfunction now that it is a one-liner.

Test:
cargo test
cargo clippy