Store RAFT_KEY_RUN in db and use for RaftContextKey
This change the serialization of RaftContextKey, deserializing when restarting should fail
for these entry, as a result they will be ignored which should not be an issue.
Ideally we would handle version change better.

Adding proposer_run ensure we will not confuse a proposal made after restart with a different
proposal made before restart.

Test:
cargo clippy
cargo test