Avoid resending messages if unrelated to requested block
If a we receive a BlockStored again before consensus is reach,
we do not want to trigger messages for the previous block to miners or storage.

Additionally, whether we mined the previous block or not, we want to
create a fresh partition with a new random number.

Note:
We generate a random number on startup to handle the case of restart
with snaphot loaded as we would not call reset function again in this case.

Test:
cargo clippy
cargo test

./src/bin/node_settings_local_raft_run.sh
./src/bin/node_settings_local_raft_run.sh no_rm