Expect Handshake Request and Response
The code was treating response as somewhate optional, allowing messages to be processed
even if the response was not recceived or malformed.

With version checking we should not be considered connected until the handshake complete,
so we error out if handshake failed, and ignore message until it is received.

Test:
cargo clippy
cargo test
src/bin/node_settings_local_raft_run.sh