Merkle tree verificaitons for miner node
The transactions from the compute node are sent by flooding to partition.
The Miner recieves this and builds the merkle tree.
Two transactions are picked at random and are checked against the tree.
If the block is valid, the block will be marked as valid and the PoW will start.
If the block is not valid, the PoW will still be calculated but it will be makred that the block in invalid.
Used NOAM functions where possible in miner.
changed transaction to tx_miner_verification
removed debug line.

Tests
Cargo clippy
cargo test
cargo build
cargo fmt -- --check