Simplify open db: always allow extra columns
That option is not necessary, after a failed upgrade, the only difference should
be the presence of the column. Also if column did not change during upgrade it
would not fail to open.

If we want to prevent db partially upgraded to open we would need to make that explicit,
and not rely on column difference.

Test:
cargo clippy
cargo test