Drop known miners on upgrade and wait for them
To mimic normal startup, we wait for the threshold of miners of the current version
to propose a block: Only when a majority of compute nodes are ready will we start
mining a block.

As a result, we drop all registered miner during upgrade. This is good because it also
take care of miner changing IP after upgrade (as happen in upgrade test).

Test:
cargo test
cargo clippy

$> rm -rf src/db/db/test.* src/wallet/wallet/test.*; tar -xzf src/bin/dbs_v_0_2_0.tar.gz
$> cargo run --bin upgrade --  --config=src/bin/node_settings_local_raft_1.toml --type all --processing upgrade
 => Success
$> src/bin/node_settings_local_raft_1_run.sh no_rm
 => Continue with next block