Add actual key to BlockchainItem
We may retreive the block/transaction with a named key (last or block index etc).,
so we also need to know the key, in particular if we are a storage node trying to
catch up and to add the blocks/transactions to the database.

Note:
We may add further metadata, in particular one could be the block number.
=> For block: don't need to deserialize to get it
=> For transaction: could tie to block currently require chain traverasl

Test:
cargo clippy
cargo test