Add simple donation process where all request are accepted
If donation are enabled, the specified config amount will be send to requested address
if we did not initiate that payment.

This simply do the second part of the payment process and can reuse all code.
New test added to validate donation, and that donation fail if not enabled.

Test:
cargo test
cargo clippy