Add library code and test to fetch raft database in pre-launch mode
We need to raft database before launching the node.
Here we are launching nodes as PreLaunchNode and the first node send its raft db to all others.
This handle a 1 to N upgrade, upgrade is necessary to update the number of node in consensused state.

Test cover upgrade same version and upgrade to new version.

Test:
cargo clippy
cargo test