Verify client name in certificate
After validating that we trust the certificate, when receiving the public address from the peer, we check that
the name associated with the address is valid for the certificate.

This provide the 2 way checks so both connecting and connected to nodes can have their identity trusted.

Test:
cargo clippy
cargo test
src/bin/node_settings_local_raft_3_run.sh