Avoid using TLS for all test (performance + noise)
When debugging with TLS enabled, a lot of info related to TLS is included and while it can be filtered
it makes it harder. Similarly when debugging this additional layer can get in the way without impacting
the logic under test.

As an additional consideration, performance is significantly impacted particularly in debug mode.
The full flow tests that are more overall validation tests mostly use TLS throughout.

Test:
cargo test
cargo clippy