Initial bring up for ring library instead of sodiumoxide for Ed25519 signing
The secret key format is taking value using from_pkcs8 function so the format needed to be updated.
We'll need to update the json file with the new format thought it might be better to use something that is produce similarly
to the key for certificate (which currently in incorrectly RSA key).

Test:
cargo clippy
cargo test