Backup syncronzied database over compute and storage regularly
We need to get backup that are consistent accross all the nodes in the network
and different databases on the same node.

This would happen with coordinated shutdown, but a similar system tend to not
work as well for backup as nodes continue to progress after backup.

Additionally, we need to take consistent backup for raft and main database.
This should happen at the time block is generated for compute
or stored for storage, when we create a snaphot, but we need to wait until the
raft snapshot is generated to take the raft database backup.

With these, it become possible to generate regular backup:

- Backup at block num that are miultiple of 1000:
backup_block_modulo = 1000

- Restore  backuped database (Danger: Will replace database, may require to remove db to restore)
backup_restore = true

Test
- cargo clippy
- cargo test
:
- src/bin/node_settings_local_raft_1_run.sh  (Run network and verify backup exists)
- src/db/db/test.*.0
- Update src/bin/node_settings_local_raft_1.toml with backup_restore = true
- src/bin/node_settings_local_raft_1_run.sh no_rm   (Run network and verify backup restored)
=> Network restarted at backuped block